refactor: division component from calendar screen, home screen
