+++ app/src/main/java/eu/tutorial/moodle/ui/component/HomeEmotion.kt
@@ -1,0 +1,65 @@
+package eu.tutorial.moodle.ui.component
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.Font
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import eu.tutorial.moodle.R
+
+@Composable
+fun EmotionComponent(
+    exist : Boolean
+){
+    if(exist){ // TODO 이 부분 코드 개선.. value를 빼서 Composable 줄이는 방향
+        Card(
+            modifier = Modifier
+                .padding(start = 12.dp, end = 12.dp),
+            shape = RoundedCornerShape(32.dp)
+        ) {
+            Box(
+                modifier = Modifier
+                    .height(300.dp)
+                    .fillMaxWidth()
+                    .background(color = Color(0XEFEFEFEF)),
+                contentAlignment = Alignment.Center
+            ){
+            }
+        }
+    }else{
+        Card(
+            modifier = Modifier
+                .padding(start = 12.dp, end = 12.dp),
+            shape = RoundedCornerShape(32.dp)
+        ) {
+            Box(
+                modifier = Modifier
+                    .height(300.dp)
+                    .fillMaxWidth()
+                    .background(color = Color(0XEFEFEFEF)),
+                contentAlignment = Alignment.Center
+            ){
+                Text(
+                    text = "Emotions",
+                    style = TextStyle(
+                        fontSize = 32.sp,
+                        fontFamily = FontFamily(Font(R.font.poppins_bold)),
+                        color = Color(0XFF414141)
+                    ),
+                )
+            }
+        }
+    }
+}
